+++ tests/Integration/Behaviour/behat.yml
@@ -66,6 +66,7 @@ default:
                 - Tests\Integration\Behaviour\Features\Context\Domain\OrderFeatureContext
                 - Tests\Integration\Behaviour\Features\Context\CarrierFeatureContext
                 - Tests\Integration\Behaviour\Features\Context\CountryFeatureContext
+                - Tests\Integration\Behaviour\Features\Context\Domain\OrderPaymentFeatureContext
 
         currency:
             paths:
